Inicio control de sesion en rutas

+++ MajorsACasa/src/main/java/es/uji/ei1027/majorsACasa/controller/BeneficiarioController.java
@@ -1,5 +1,7 @@
 package es.uji.ei1027.majorsACasa.controller;
 
+import javax.servlet.http.HttpSession;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataAccessException;
 import org.springframework.dao.DuplicateKeyException;
@@ -13,6 +15,7 @@ import org.springframework.web.bind.annotation.RequestMethod;
 
 import es.uji.ei1027.majorsACasa.dao.BeneficiarioDao;
 import es.uji.ei1027.majorsACasa.model.Beneficiario;
+import es.uji.ei1027.majorsACasa.model.UserDetails;
 
 @Controller
 @RequestMapping("/beneficiario")
@@ -26,13 +29,25 @@ public class BeneficiarioController {
 	}
 
 	@RequestMapping("/list")
-	public String listBeneficiarios(Model model) {
+	public String listBeneficiarios(HttpSession session, Model model) {
+		UserDetails user = (UserDetails)session.getAttribute("user");
+	    if (user == null || !(user.getTipo().equals("jefe"))) { 
+          model.addAttribute("user", new UserDetails());
+          session.setAttribute("nextUrl", "beneficiario/list");
+          return "login";
+        }
 		model.addAttribute("beneficiarios", beneficiarioDao.getBeneficiarios());
 		return "beneficiario/list";
 	}
 	
 	@RequestMapping(value="add")
-	public String addBeneficiario(Model model) {
+	public String addBeneficiario(HttpSession session, Model model) {
+		UserDetails user = (UserDetails)session.getAttribute("user");
+	    if (user == null || !(user.getTipo().equals("jefe"))) { 
+          model.addAttribute("user", new UserDetails());
+          session.setAttribute("nextUrl", "beneficiario/add");
+          return "login";
+        }
 		model.addAttribute("beneficiario", new Beneficiario());
 		return "beneficiario/add";
 	}
@@ -57,7 +72,13 @@ public class BeneficiarioController {
 	}
 	
 	@RequestMapping(value="/update/{id}", method = RequestMethod.GET) 
-	public String editBeneficiario(Model model, @PathVariable String id) { 
+	public String editBeneficiario(HttpSession session, Model model, @PathVariable String id) { 
+		UserDetails user = (UserDetails)session.getAttribute("user");
+	    if (user == null || !(user.getTipo().equals("jefe") || user.getTipo().equals("beneficiario"))) { 
+          model.addAttribute("user", new UserDetails());
+          session.setAttribute("nextUrl", "beneficiario/update");
+          return "login";
+        }
 		model.addAttribute("beneficiario", beneficiarioDao.getBeneficiario(id));
 		return "beneficiario/update"; 
 	}
@@ -74,7 +95,13 @@ public class BeneficiarioController {
 	}
 	
 	@RequestMapping(value="/delete/{id}")
-	public String processDelete(@PathVariable String id) {
+	public String processDelete(HttpSession session, Model model, @PathVariable String id) {
+		UserDetails user = (UserDetails)session.getAttribute("user");
+	    if (user == null || !(user.getTipo().equals("jefe") || user.getTipo().equals("beneficiario"))) { 
+          model.addAttribute("user", new UserDetails());
+          session.setAttribute("nextUrl", "redirect:../list");
+          return "login";
+        }
 		beneficiarioDao.deleteBeneficiario(id);
 		return "redirect:../list";
 	}
